Don't render until target is ready inside target window.
This prevents issues with key handlers in <input> and <textarea> elements. It
appears that React sets up a number of event proxies on mount, and if the
element is first mounted in one document then translocated to another,
these proxies don't follow. So we resist mounting at all until the
child document is ready.

Added a story mirroring text input to illustrate this. It will
fail on the old version.